// src/app/models/admin/hosogiayto/cptailieu.model.ts
export class OutputCpTaiLieuModel {
  idtailieu: string;
  tentailieu: string;
  filedinhkem: string;
  duongdan: string;
  thutu: number;
  dungluong: number;
  dinhdang: string;
  idgiayphep: string;
  siteid: string;
}

export class InputCpTaiLieuModel {
  idtailieu: string;
  tentailieu: string;
  filedinhkem: string;
  duongdan: string;
  thutu: number;
  dungluong: number;
  dinhdang: string;
  idgiayphep: string;
  siteid: string;
}